requirements.txt
ccxt>=4.3.0
python-dotenv>=1.0.1
python-telegram-bot>=21.0
